Create lint file, fixes moezbhatti/qksms#658
There is some ProGuard issue with inner classes in one of the libraries
that we're using in QKSMS. One warning is written as 8 errors in the
linter, and there's like 98 warnings (one for each class), so we get
700+ errors. The only consequence of ignoring the warning is this:

    The consequence of ignoring this warning is that reflective
    operations on this class will incorrectly indicate that it is *not*
    an inner class. [From the warning message itself.]

It doesn't seem likely that QKSMS is going to break because the classes
in a library are being reported as not an inner class during reflection.
So I've disabled the printing of the errors with a lint.xml file.

You can read more about disabling error messages here:
http://tools.android.com/tips/lint/suppressing-lint-warnings